Refactor: Sidebar 구조 리팩토링
